requirements.txt
aiogram>=3.7.0
fastapi
uvicorn
python-dotenv
sqlalchemy
aiosqlite
aiofiles
python-multipart
